Descomponentizo el nav. Agrego texto a botones a partir de tablet. Modifico hoja de estilo.
